Field alter change added in already existing migration file

+++ infraohjelmointi_api/models/Project.py
@@ -17,7 +17,7 @@ class Project(models.Model):
     siteId = models.ForeignKey(
         BudgetItem, on_delete=models.DO_NOTHING, null=True, blank=True
     )
-    hkrId = models.IntegerField(blank=True, null=True)
+    hkrId = models.PositiveBigIntegerField(blank=True, null=True)
     entityName = models.CharField(max_length=30, blank=True, null=True)
     sapProject = models.CharField(max_length=100, blank=True, null=True)
     sapNetwork = models.JSONField(blank=True, null=True)
